Implement keyboard navigation
This commit adds a toggle to the settings menu to enable keyboard navigation. The toggle adds a focus highlighter to buttons. Keyboard navigation only applies to UI elements on title screen and quit button on battle scene. Future updates will fine tune appearance of highlighter and expand access to other parts of the game.