MapGenerator functionality added
The MapGenerator function will create a 2D list of map_space nodes with instanced variables set, including position in the grid (with offsets) and connections from the starting nodes to the boss nodes. There are no visuals currently for the player to interact with, however running the scene will print a debug message to see which nodes are active (part of the path) and what type of node they are.