Run gameloop
Created the Run scene which will be the main game loop. Run has 2 nodes: Overworld, which has the map and all of the functionality for interacting with the map, and another node which gets instantiated with whatever scene is required (battle, shop, event, treasure). When a node is clicked, the map is hidden and the instantiated scene is shown, when that scene completes it will show the map again.